fix: check if beetsDB file exists before unlink

+++ src-backend/controllers/metaController.js
@@ -148,8 +148,10 @@ export const writeMetaToFile = (req, res) => {
   checkFileExists(filePath, fileName, res);
 
   // Remove beets db file to avoid issues
-  fs.unlink(beetsDB);
-
+  if (fs.existsSync(beetsDB)) {
+    fs.unlink(beetsDB);
+  }
+  
   const beetsArgs = [
     'import',
     '--write',
